requirements.txt
django==1.9
coverage==4.0.3
coveralls==1.1
djangorestframework==3.3.2
